Refactor button mechanism and LCD UI for improved functionality and debugging, also create code documentation

+++ projects/9. Robotic Arm/include/lcd_ui.h
@@ -3,17 +3,19 @@
 
 #include <LiquidCrystal_I2C.h>
 #include <ToneESP32.h>
-
 #include <pin_definition.h>
 
 
 
 //* Variable Declaration
+
+// Enum for robot/servo mode
 enum SystemMode{MODE_MANUAL,
                 MODE_AUTO_1,
                 MODE_AUTO_2,
                 MODE_WIFI};
 
+// 
 enum menuState{
                 OF_SETTING,ON_SETTING
             };
@@ -23,7 +25,7 @@ extern String menuItem[];
 extern uint64_t lcdTimer;
 extern uint64_t menuStartTimer;
 extern uint_fast8_t menuTimeout;
-extern uint_fast8_t blink;
+extern uint_fast64_t blink;
 extern uint_fast8_t blinkInterval;
 
 extern SystemMode currentMode;
